refactor : make independent 'PlaceSearch' component and refactor with ImperatvieHandle
