+++ src/components/ui/SearchBar/style.ts
